requirements.txt
alabaster==0.7.13
Babel==2.12.1
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
docutils==0.18.1
idna==3.4
imagesize==1.4.1
importlib-metadata==6.3.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
pipfile==0.0.2
Pygments==2.14.0
python-dotenv==1.0.0
requests==2.28.2
snowballstemmer==2.2.0
Sphinx==6.1.3
sphinx-rtd-theme==1.2.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
toml==0.10.2
urllib3==1.26.15
zipp==3.15.0
